changes for MinGW compatibility has to cast parameters to LONGS_EQUAL to (long) so that eclipse does not show it as an error. Changed  TEST(MockFailureTest, MockUnexpectedObjectFailure) so that differences in how %p is handled on MinGW do not cause test failures
Added TEST(SimpleString, StringFromFormatpointer) to catch %p problems in the future

+++ include/CppUTest/UtestMacros.h
@@ -143,7 +143,7 @@
   LONGS_EQUAL_LOCATION(expected,actual,__FILE__, __LINE__)
 
 #define LONGS_EQUAL_LOCATION(expected,actual,file,line)\
-  { Utest::getCurrent()->assertLongsEqual(expected, actual,  file, line); }
+  { Utest::getCurrent()->assertLongsEqual((long)expected, (long)actual,  file, line); }
 
 #define BYTES_EQUAL(expected, actual)\
     LONGS_EQUAL((expected) & 0xff,(actual) & 0xff)
